pyproject.toml
[project]
name = "route-proxy"
version = "0.1.0"
description = "Lightweight proxy: Responses API <-> Chat Completions API with forced streaming"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
